refactor(utils): move memoize to its own file
